Project token pallet: generate_storage_info

+++ runtime-modules/content/src/tests/mock.rs
@@ -1000,7 +1000,7 @@ impl common::working_group::WorkingGroupBudgetHandler<U256, u64> for Distributio
 // pallet_project_token trait implementation and related stuff
 parameter_types! {
     pub const TokenModuleId: PalletId = PalletId(*b"m__Token");
-    pub const MaxVestingSchedulesPerAccountPerToken: u8 = 3;
+    pub const MaxVestingSchedulesPerAccountPerToken: u32 = 3;
     pub const BlocksPerYear: u32 = 5259487; // blocks every 6s
 }
